pyproject.toml
[tool.poetry]
name = "tortoiesorm-auth-microservice"
version = "0.1.0"
description = ""
authors = ["vladimirmac <vladimirbodrin89@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi-debug-toolbar = "^0.6.2"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
python-dotenv = "^1.0.1"
tortoise-orm = "^0.20.0"
pydantic = {version = "2.6", extras = ["email"]}
aerich = "^0.7.2"
asyncpg = "^0.29.0"
grpcio-tools = "^1.62.2"


[tool.poetry.group.dev.dependencies]
pytest-factoryboy = "^2.7.0"
httpx = "^0.27.0"
pre-commit = "^3.7.0"
ruff = "^0.4.1"


[tool.aerich]
tortoise_orm = "src.config.settings.DATABASE_CONFIG"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
